cleaned proj code, and progress on auth

+++ projectroot/src/main/java/com/cebedo/pmsys/systemuser/service/SystemUserServiceImpl.java
@@ -5,6 +5,9 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.cebedo.pmsys.common.AuthUtils;
+import com.cebedo.pmsys.company.model.Company;
+import com.cebedo.pmsys.login.authentication.AuthenticationToken;
 import com.cebedo.pmsys.systemuser.dao.SystemUserDAO;
 import com.cebedo.pmsys.systemuser.model.SystemUser;
 
@@ -21,30 +24,49 @@ public class SystemUserServiceImpl implements SystemUserService {
 	@Transactional
 	public void create(SystemUser systemUser) {
 		this.systemUserDAO.create(systemUser);
+		AuthenticationToken auth = AuthUtils.getAuth();
+		Company authCompany = auth.getCompany();
+		if (AuthUtils.notNullObjNotSuperAdmin(authCompany)) {
+			systemUser.setCompany(authCompany);
+			this.systemUserDAO.update(systemUser);
+		}
 	}
 
 	@Override
 	@Transactional
 	public SystemUser getByID(long id) {
-		return this.systemUserDAO.getByID(id);
+		SystemUser obj = this.systemUserDAO.getByID(id);
+		if (AuthUtils.isActionAuthorized(obj)) {
+			return obj;
+		}
+		return new SystemUser();
 	}
 
 	@Override
 	@Transactional
 	public void update(SystemUser systemUser) {
-		this.systemUserDAO.update(systemUser);
+		if (AuthUtils.isActionAuthorized(systemUser)) {
+			this.systemUserDAO.update(systemUser);
+		}
 	}
 
 	@Override
 	@Transactional
 	public void delete(long id) {
-		this.systemUserDAO.delete(id);
+		SystemUser obj = this.systemUserDAO.getByID(id);
+		if (AuthUtils.isActionAuthorized(obj)) {
+			this.systemUserDAO.delete(id);
+		}
 	}
 
 	@Override
 	@Transactional
 	public List<SystemUser> list() {
-		return this.systemUserDAO.list();
+		AuthenticationToken token = AuthUtils.getAuth();
+		if (token.isSuperAdmin()) {
+			return this.systemUserDAO.list(null);
+		}
+		return this.systemUserDAO.list(token.getCompany().getId());
 	}
 
 	@Override
